Fix an error where the lists of processes are inconsistent
In some situation where the list of processes obtained from
`nvidia-smi` (gpu usage query) and `ps aux` commands are inconsistent,
an exception (TypeError: 'NoneType' object is not iterable) occurs.

This is because in some version of Nvidia drivers or in other environments,
nvidia-smi might return invalid processes that are not currently running
in the OS. For a workaround, we ignore such entries.

Fixes #1.