Introduce --color option to force colored output
With introducing blessings module (from 0.4) which automatically
determines color for output, depending on whether stdout has tty
or not (i.e. piped), we should have a way to 'force' colored output.

For example, `watch --color gpustat` was always suppressing color.
Using the `--color` or `--force-color` option, we can always have
the colored output even if stdout is not a tty.