Ignore FileNotFoundError while querying process information (#95)
It appears that there is a bug of psutil where FileNotFoundError
exceptions are not properly translated to NoSuchProcess when the
requested process actually does not exist. As a workaround, we ignore
FileNotFoundError exceptions thrown while querying process information
as well as NoSuchProcess.

We also require psutil to have a minimum version of 5.6.0+ to ensure
that the bug (hopefully and probablyj) does not happen.