Use NVIDIA's official pynvml binding
Since 2021, NVIDIA provies an official python binding `pynvml`,
which should replace existing third-party (not maintained) one.
The APIs are mostly the same, but it ships with some recent features
such as MIG support.

To ensure a correct `pynvml` package is installed, we perform
some sanity check upon importing the library.