Make gpustat.nvml compatible with a third-party fork of pynvml (#153)
An un-official distribution of the pynvml package provided by
@gpuopenanalytics/pynvml was causing an incompatibility issue, where
gpustat.nvml fails to import. Although the official and correct pynvml
packages should be installed, but wild users' python environments can be
easily messed up by unawaringly installing the problematic, unofficial
pynvml package.

gpustat can support this third-party pynvml by directly importing the
low-level pynvml.nvml package (which should be exactly the same as
an official pynvml.py). This is a quickfix that relies on the
implementation detail of a third-party package, but nevertheless it
should be a good enough workaround solution for many users in practice.