Bump up nvidia-ml-py requirement to 12.535.108
nvidia-ml-py==12.535.77 is a buggy version that breaks the struct for
process information, and should not be used (unless NVIDIA driver is
*also* buggy, 535.43, 535.54, and 535.86). The latest version
nvidia-ml-py==12.535.108 fixes the problem and is still compatible with
our supported drivers (R450+).

To ensure users who will install gpustat 1.2.0 have a correct version
of nvidia-ml-py version installed, we bump up the requirement.

See #160 and #161 for more details.